Add float tests and pull instruction tests into separate modules
